incorporated view oqs.query_stats
The view oqs.query_stats is a wrapper for sys.dm_exec_query_stats

This allows us to have dummy fields for lower versions of SQL Server which don't have the funky new columns. The dynamic SQL (yes, I feel dirty), recognises the version of SQL Server and inserts dummy fields into the view as necessary.

The Gather_Statistics sproc now references this view so that we can be sure to have all the columns regardless of version. If a dummy column is present it either returns a NULL or a 0 value, depending on data type and usage of that column.