Create uninstall_open_query_store_non_db_items.sql
The uninstall_open_query_store.sql will remove all objects from an instance. But when an installation on a second database is unsuccesful, also objects like the login, certificate, agent job and startup procedure are removed from the instance.
This uninstall file will only remove the objects, outside of the database. So it can be used if the initial install in the instance is unsuccessful.