Update uninstall_open_query_store.sql
When a second database in the same instance is configured for OQS but the installation fails, all objects are removed from the instance. Also objects used by the first installed database.
To prevent this, the uninstall_open_query_store. sql can be split into 2 parts, one for the database items, one for the objects on instance level. 
I also added a file as a pull request, with the uninstallation of all instance-related objects.